comienzo de logica de paralelizacion
Comienzo de creacion de abstract client y limpieza en server
