--- protocolo.md
@@ -1,7 +1,0 @@
-### HEADER
-|##|#|
-sizeof|last|
-
-
-#### sizeof
-size of full message
